unplatform_source/requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports-abc==0.4
billiard==3.3.0.23
celery==3.1.17
dj-database-url==0.4.0
Django==1.8.6
django-cors-headers==1.1.0
django-rest-swagger==0.3.7
djangorestframework==3.3.2
kombu==3.0.35
maproxy==0.0.12
MarkupSafe==0.23
PyInstaller==3.2.0
pytz==2016.3
PyYAML==3.11
requests==2.12.1
tornado==4.3
waitress==0.8.10
natsort==5.0.1
